Add STM finalizers
This introduces a new function

   atomicallyWithIO :: STM a -> (a -> IO b) -> IO b

Like the existing atomically operation, atomicallyWithIO transforms a
computation of type STM a into an I/O action. Additionally, it takes a
finalizer, which is a function of type a -> IO b, viz. an I/O action
that can depend on the result of the STM computation, and combines it
with the transaction in such a way that:

  1. The finalizer is only performed if the STM transaction is
     guaranteed to commit.

  2. The STM transaction only commits (i.e. makes its effects visible
     to other transactions) if the finalizer finishes without raising
     an exception.

A detailed specification, including an operational semantics, as well
as a discussion of the implementation, can be found in chapter 2 of my
master's thesis (http://github.com/mcschroeder/thesis).